change console app class to KusanagiBot.cs
